feat: target architecture and profile are optional arguments to the `build` command
- Removed redundant types from cargo-wdk, using types from wdk-build itself
- Implemented FromStr and fmt::Display traits for the CpuArchitecture and DriverConfig
- Fixed tests
