+++ crates/cargo-wdk/src/actions/package/package_task.rs
@@ -14,10 +14,9 @@ use std::{
 
 use mockall_double::double;
 use tracing::{debug, info};
-use wdk_build::DriverConfig;
+use wdk_build::{CpuArchitecture, DriverConfig};
 
 use super::error::PackageTaskError;
-use crate::actions::CpuArchitecture;
 #[double]
 use crate::providers::{exec::CommandExec, fs::Fs, wdk_build::WdkBuild};
